+++ tests/test_tiling.py
@@ -1,5 +1,5 @@
 import json
-from itertools import chain
+from itertools import chain, product
 
 import pytest
 
@@ -675,8 +675,7 @@ def test_cells_in_row_col(typical_redundant_obstructions,
     assert col4 == set()
 
 
-def test_cell_basis(typical_redundant_obstructions,
-                    typical_redundant_requirements):
+def test_cell_basis(factorable_tiling):
     tiling = Tiling(
         obstructions=[Obstruction(Perm((0, 2, 1)), [(0, 0), (0, 0), (0, 0)]),
                       Obstruction(Perm((0, 2, 1)), [(0, 0), (0, 1), (1, 1)]),
@@ -698,10 +697,15 @@ def test_cell_basis(typical_redundant_obstructions,
     basis = bdict[(1, 1)]
     assert len(basis[1]) == 0
     assert set(basis[0]) == {Perm((0, 2, 1))}
-    basis = bdict[(3, 3)]
-    assert len(basis) == 2
-    assert basis[0] == []
-    assert basis[1] == []
+    # Basis for a non active cell
+    bdict = factorable_tiling.cell_basis()
+    assert bdict[(0, 1)] == ([Perm((0,))], [])
+    # Check that all cell have a basis
+    dim = factorable_tiling.dimensions
+    for cell in product(range(dim[0]), range(dim[1])):
+        assert cell in bdict
+        assert len(bdict[cell][0]) >= 1
+    assert dim + (1, 0) not in bdict
 
 
 def test_cell_graph(factorable_tiling, compresstil,